# servicios-sociales-api/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /api/users:
    post:
      operationId: UsersController_create
      parameters: []
      responses:
        '201':
          description: ''
      security:
        - JWT-auth: []
      summary: Registrar un nuevo usuario (Personal Interno)
      tags:
        - users
    get:
      operationId: UsersController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
      security:
        - JWT-auth: []
      summary: Listar usuarios del sistema (Personal)
      tags:
        - users
  /api/users/{id}:
    get:
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      security:
        - JWT-auth: []
      summary: Obtener un usuario por ID
      tags:
        - users
    patch:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - JWT-auth: []
      summary: Actualizar datos básicos de un usuario
      tags:
        - users
    delete:
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - JWT-auth: []
      summary: Eliminar un usuario del sistema
      tags:
        - users
...
  /api/medical-record:
    post:
      operationId: MedicalRecordController_create
      parameters: []
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
      security:
        - JWT-auth: []
      summary: Registrar una nueva consulta médica
      tags:
        - medical-record
    get:
      operationId: MedicalRecordController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
      security:
        - JWT-auth: []
      summary: Listar todos los expedientes
      tags:
        - medical-record
  /api/medical-record/{id}:
    get:
      operationId: MedicalRecordController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      security:
        - JWT-auth: []
      summary: Obtener un expediente por ID
      tags:
        - medical-record
  /api/medical-record/citizen/{citizenId}:
    get:
      operationId: MedicalRecordController_findByCitizenId
      parameters:
        - name: citizenId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - JWT-auth: []
      summary: Obtener expedientes de un ciudadano específico
      tags:
        - medical-record
...
info:
  title: Servicios Sociales API
  description: |-
    API de autenticación y gestión para instituciones de servicios sociales de El Salvador.
          
    **Instituciones soportadas:** ISSS, AFP, ISBM, IPSFA, INPEP

    **Autenticación:** JWT Bearer Token (RS256)

    **Roles disponibles:** super_admin, doctor, institution_staff, citizen

    **Credenciales de prueba:**
    - Super Admin: admin@ssapi.gob.sv / Admin123!
    - Ciudadanos: Usar datos generados en la tabla Citizen (ej. juan.perez@example.com)
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Ingresa tu JWT access token
      in: header
...